Initial setup
This setup contains a fully working FastAPI server with a base
implementation of the Node RED interface for NXT interaction.

It currently allows to add a mock motor block and allows to
connect (typed) ports to each other. Some mock controls have been
added.

This state is an early prototype to make sure the fundamentals
work right.